Update to avoid breaking other vendor backends.
