+++ include/spblas/detail/view_inspectors.hpp
@@ -84,13 +84,20 @@ bool has_scaling_factor(T&& t) {
 template <tensor T>
 auto get_ultimate_base(T&& t) {
   if constexpr (has_base<T>) {
-    if constexpr (is_matrix_opt_view_v<T>) {
-      return t;
-    } else {
-      return get_ultimate_base(t.base());
-    }
+    return get_ultimate_base(t.base());
   } else {
     return t;
+  }
+}
+
+template <tensor T>
+bool has_matrix_opt(T&& t) {
+  if constexpr (is_matrix_opt_v<T>) {
+    return true;
+  } else if constexpr (has_base<T>) {
+    return has_matrix_opt(t.base());
+  } else {
+    return false;
   }
 }
 
@@ -98,17 +105,13 @@ template <typename T>
 using ultimate_base_type_t = decltype(get_ultimate_base(std::declval<T>()));
 
 template <typename T>
-concept has_csr_base =
-    is_csr_view_v<ultimate_base_type_t<T>> ||
-    (is_matrix_opt_view_v<ultimate_base_type_t<T>> &&
-     is_csr_view_v<decltype(std::declval<ultimate_base_type_t<T>>().base())>);
+concept has_csr_base = is_csr_view_v<ultimate_base_type_t<T>>;
 
 template <typename T>
 concept has_csc_base = is_csc_view_v<ultimate_base_type_t<T>>;
 
 template <typename T>
-concept has_mdspan_matrix_base =
-    is_matrix_instantiation_of_mdspan_v<ultimate_base_type_t<T>>;
+concept has_mdspan_matrix_base = is_matrix_mdspan_v<ultimate_base_type_t<T>>;
 
 template <typename T>
 concept has_contiguous_range_base =
