+++ include/spblas/views/inspectors.hpp
@@ -3,6 +3,7 @@
 #include <spblas/detail/mdspan.hpp>
 #include <spblas/views/csc_view.hpp>
 #include <spblas/views/csr_view.hpp>
+#include <spblas/views/matrix_opt.hpp>
 #include <spblas/views/scaled_view.hpp>
 
 namespace spblas {
@@ -57,8 +58,40 @@ struct is_matrix_instantiation_of_mdspan<
 };
 
 template <typename T>
-static constexpr bool is_matrix_instantiation_of_mdspan_v =
+static constexpr bool is_matrix_mdspan_v =
     is_matrix_instantiation_of_mdspan<std::remove_cvref_t<T>>::value;
+
+template <typename T>
+struct is_instantiation_of_scaled_view {
+  static constexpr bool value = false;
+};
+
+template <typename S, typename T>
+struct is_instantiation_of_scaled_view<scaled_view<S, T>> {
+  static constexpr bool value = true;
+};
+
+template <typename T>
+static constexpr bool is_scaled_view_v =
+    is_instantiation_of_scaled_view<std::remove_cvref_t<T>>::value;
+
+template <typename T>
+static constexpr bool is_scaled_view_matrix_v =
+    is_scaled_view_v<T> && matrix<decltype(std::declval<T>().base())>;
+
+template <typename T>
+struct is_instantiation_of_matrix_opt {
+  static constexpr bool value = false;
+};
+
+template <typename T>
+struct is_instantiation_of_matrix_opt<matrix_opt<T>> {
+  static constexpr bool value = true;
+};
+
+template <typename T>
+static constexpr bool is_matrix_opt_v =
+    is_instantiation_of_matrix_opt<std::remove_cvref_t<T>>::value;
 
 } // namespace __detail
