+++ include/spblas/views/views.hpp
@@ -4,6 +4,7 @@
 #include <spblas/views/csr_view.hpp>
 #include <spblas/views/inspectors.hpp>
 #include <spblas/views/matrix_opt.hpp>
+#include <spblas/views/matrix_opt_impl.hpp>
 #include <spblas/views/scaled_view.hpp>
 #include <spblas/views/scaled_view_impl.hpp>
 #include <spblas/views/view_base.hpp>
